Update display.yaml
Is this going to be in the next public release? I made a few wording improvements, but this import method needs better documentation.

What goes in the TSV file? Is there an example somewhere?

What does "Flag for filling in missing values in matrix" do? The long-hint needs more info.

What does "Scale of values in matrix" mean, and how should it be represented? Should it be a range, like "0-1"?